requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements_dev.in
#
aiohttp==3.9.5
    # via
    #   -r requirements_dev.in
    #   aioresponses
    #   pytest-aiohttp
aioresponses==0.7.6
    # via -r requirements_dev.in
aiosignal==1.3.1
    # via
    #   -r requirements_dev.in
    #   aiohttp
asyncio==3.4.3
    # via -r requirements_dev.in
attrs==23.2.0
    # via
    #   -r requirements_dev.in
    #   aiohttp
coverage[toml]==7.5.1
    # via
    #   -r requirements_dev.in
    #   pytest-cov
dateutils==0.6.12
    # via -r requirements_dev.in
frozenlist==1.4.1
    # via
    #   -r requirements_dev.in
    #   aiohttp
    #   aiosignal
idna==3.7
    # via
    #   -r requirements_dev.in
    #   yarl
iniconfig==2.0.0
    # via
    #   -r requirements_dev.in
    #   pytest
logger==1.4
    # via -r requirements_dev.in
multidict==6.0.5
    # via
    #   -r requirements_dev.in
    #   aiohttp
    #   yarl
orjson==3.10.3
    # via -r requirements_dev.in
packaging==24.0
    # via
    #   -r requirements_dev.in
    #   pytest
pluggy==1.5.0
    # via
    #   -r requirements_dev.in
    #   pytest
pytest==8.2.0
    # via
    #   -r requirements_dev.in
    #   pytest-aiohttp
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
pytest-aiohttp==1.0.5
    # via -r requirements_dev.in
pytest-async==0.1.1
    # via -r requirements_dev.in
pytest-asyncio==0.23.6
    # via
    #   -r requirements_dev.in
    #   pytest-aiohttp
pytest-cov==5.0.0
    # via -r requirements_dev.in
pytest-mock==3.14.0
    # via -r requirements_dev.in
python-dateutil==2.9.0.post0
    # via dateutils
pytz==2024.1
    # via dateutils
six==1.16.0
    # via python-dateutil
strenum==0.4.15
    # via -r requirements_dev.in
typing==3.7.4.3
    # via -r requirements_dev.in
urllib3==2.2.1
    # via -r requirements_dev.in
yarl==1.9.4
    # via
    #   -r requirements_dev.in
    #   aiohttp
